+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Added
 * Changelog
 * ``interp_bounds`` formatoption for the ``plot2d`` plot method (see the
   `docs <https://psyplot.readthedocs.io/projects/psy-simple/en/latest/api/psy_simple.plotters.html#psy_simple.plotters.Simple2DPlotter.interp_bounds>`__)
+* Added the ``fldmean`` plot method that can be used to directly calculate and
+  plot the mean over the x- and y-dimensions
 
 Changed
 -------